% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions_ramses.R
\name{dmnorm}
\alias{dmnorm}
\title{Multivariate Normal Density Function}
\usage{
dmnorm(y, mu, cov, prec, log = FALSE)
}
\arguments{
\item{y}{vector of values.}

\item{mu}{mean vector.}

\item{cov}{covariance matrix.}

\item{prec}{precision matrix.}

\item{log}{logical; if TRUE, density calculations are computed on the log scale.}
}
\description{
Computes the (log) density of the multivariate normal distribution
using either the covariance or precision parametrization.
}
