% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chart_builder.R
\name{mod_chart_builder_server}
\alias{mod_chart_builder_server}
\title{Logique serveur pour le module créateur de graphiques}
\usage{
mod_chart_builder_server(id, data_holder, append_to_rmd)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}

\item{data_holder}{Réactif contenant le dataframe actif (\code{data_holder$df}) et son nom (\code{data_holder$name}).}

\item{append_to_rmd}{Fonction d'ajout au Journal R Markdown.}
}
\value{
Un module serveur Shiny.
}
\description{
Gère l'interactivité du créateur graphique, la validation des variables,
la construction dynamique du code ggplot2 et le rendu interactif Plotly.
}
