% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chart_builder.R
\name{mod_chart_builder_ui}
\alias{mod_chart_builder_ui}
\title{Interface utilisateur pour le module créateur de graphiques}
\usage{
mod_chart_builder_ui(id)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
}
\value{
Interface utilisateur Shiny (\code{tagList}).
}
\description{
Module dédié à la création graphique interactive et intuitive
dans le package Ramses. Inspiré de l'ergonomie de Tableau Software, il permet
de sélectionner des variables réparties par type (numériques vs facteurs),
de mapper les dimensions esthétiques (X, Y, Couleur, Taille, Facet), de choisir
parmi 8 types de visualisations ggplot2 et de personnaliser les thèmes, palettes,
titres et étiquettes. Le graphique est rendu de manière interactive via Plotly
et le code ggplot2 reproductible est affiché en direct et injectable dans le Journal R Markdown.
}
