% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_prep.R
\name{mod_data_prep_server}
\alias{mod_data_prep_server}
\title{Logique serveur du module de preparation des donnees}
\usage{
mod_data_prep_server(id, data_holder, append_to_rmd)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
\item{data_holder}{Objet reactiveValues contenant le dataset actif.}
\item{append_to_rmd}{Fonction de rappel pour consigner le code R genere dans le journal R Markdown.}
}
\value{
NULL (gestion reactive Shiny).
}
\description{
Gere les interactions, les apercus avant/apres, le moteur de pipeline, l'annulation
et la consignation automatique du code R reproductible de preparation.
}
