% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_prep.R
\name{mod_data_prep_ui}
\alias{mod_data_prep_ui}
\title{Interface utilisateur du module de preparation des donnees}
\usage{
mod_data_prep_ui(id)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
}
\value{
Interface utilisateur Shiny (layout bslib).
}
\description{
Construit l'interface du module de preparation et nettoyage des donnees de Ramses,
organisee en 4 onglets : Explorer, Modifier les variables, Nettoyer, et Filtrer & Trier.
}
