% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_descriptives.R
\name{mod_descriptives_server}
\alias{mod_descriptives_server}
\title{Logique serveur pour le module de statistiques descriptives}
\usage{
mod_descriptives_server(id, data_holder, append_to_rmd)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}

\item{data_holder}{Environnement ou liste réactive contenant \code{df} et \code{name}.}

\item{append_to_rmd}{Fonction de callback pour enregistrer le code dans le journal Rmd.}
}
\value{
Un module serveur Shiny.
}
\description{
Gère la logique interactive, les calculs de statistiques descriptives,
le rendu des tableaux DT, la génération des graphiques Plotly et l'enregistrement
reproductible dans le journal R Markdown.
}
