% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_descriptives.R
\name{mod_descriptives_ui}
\alias{mod_descriptives_ui}
\title{Interface utilisateur pour le module de statistiques descriptives}
\usage{
mod_descriptives_ui(id)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
}
\value{
Un objet tagList d'interface Shiny (\code{shiny.tag}).
}
\description{
Construit l'interface du module de statistiques descriptives de Ramses,
comprenant l'analyse univariée quantitative, l'analyse univariée qualitative,
les tableaux de contingence croisés et les matrices de corrélation.
}
