% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tests.R
\encoding{UTF-8}
\name{mod_tests_prop_ui}
\alias{mod_tests_prop_ui}
\title{Interface utilisateur pour les tests de proportions dans Ramses}
\usage{
mod_tests_prop_ui(id)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
}
\value{
Un objet tagList d'interface Shiny (\code{shiny.tag}).
}
\description{
Construit l'interface dédiée aux tests de proportions (test d'une proportion vs p0
et comparaison de deux proportions indépendantes).
}
