% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tests.R
\name{mod_tests_server}
\alias{mod_tests_server}
\title{Logique serveur pour le module de tests statistiques et modélisation}
\usage{
mod_tests_server(id, data_holder, append_to_rmd)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}

\item{data_holder}{reactiveValues contenant \code{df} et \code{name}.}

\item{append_to_rmd}{Fonction de rappel pour injecter le code dans le journal Rmd.}
}
\value{
Un module serveur Shiny.
}
\description{
Exécute les tests statistiques d'hypothèse paramétriques et
non paramétriques, génère les sorties statistiques formatées, produit
les visualisations diagnostiques Plotly et consigne les commandes reproductibles
dans le journal R Markdown.
}
