% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tests.R
\name{mod_tests_ui}
\alias{mod_tests_ui}
\title{Interface utilisateur pour le module de tests statistiques et modélisation}
\usage{
mod_tests_ui(id)
}
\arguments{
\item{id}{Identifiant de namespace Shiny.}
}
\value{
Un objet tagList d'interface Shiny (\code{shiny.tag}).
}
\description{
Construit l'interface utilisateur pour les tests d'inférence statistique
et de modélisation (normalité, comparaison 2 groupes, ANOVA/Kruskal-Wallis,
tests d'indépendance du Chi-2 et Fisher, corrélations, régression linéaire et logistique).
}
