\name{ramses-package}
\alias{ramses-package}
\alias{ramses}
\docType{package}
\title{
\packageTitle{ramses}
}
\description{
\packageDescription{ramses}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ramses}
\packageIndices{ramses}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ramses}

Maintainer: \packageMaintainer{ramses}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
