% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions_ramses.R
\name{sampler}
\alias{sampler}
\title{RAMSES Sampling Structure}
\usage{
sampler(inits, fixed, updates, niter, nburn)
}
\arguments{
\item{inits}{list of named parameter initial values.}

\item{fixed}{list of named values that remain fixed in the model.}

\item{updates}{list of update functions with names matching those in parm_state.}

\item{niter, nburn}{number of MCMC iterates and burnin.}
}
\value{
A matrix where each row is an MCMC iterate and each column is a
  parameter or part of a parameter vector.
}
\description{
Function that fits a heirarchical model using MCMC. It requires a list of initial
parameter values, a list of needed fixed values, and a list of update functions,
all of which are tied together by matching element names.
}
