% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_ramses.R
\name{subject_subset}
\alias{subject_subset}
\title{Subsetting Vector By Subject ID Number}
\usage{
subject_subset(sub_id, total_sub, vec)
}
\arguments{
\item{sub_id}{and integer specifying what subject is being considered.}

\item{total_sub}{the total number of subjects accounted for in the vector.}

\item{vec}{vector to be subsetted.}
}
\description{
Generates a vector used for subsetting given a subject number, 
the total number of subjects, and the length of the vector to be subsetted.
}
