% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions_ramses.R
\name{update_step}
\alias{update_step}
\title{Update Step Routine}
\usage{
update_step(parm_state, fixed, updates, parm_names = names(updates))
}
\arguments{
\item{parm_state}{list that names parameters and contains current values.}

\item{fixed}{list of named values that remain fixed in the model.}

\item{updates}{list of update functions with names matching those in parm_state.}

\item{parm_names}{vector naming the parameters as the appear in order in both updates and parm_state.}
}
\value{
A list in the form of parm_state, but with updated values. To be used within the RAMSES sampler function.
}
\description{
Excecutes one full cycle of an MCMC iteration within the RAMSES sampling structure, updating each parameter in the parm_state list with the update function found in the updates list.
}
