% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions_ramses.R
\name{update_metropolis}
\alias{update_metropolis}
\title{Metropolis-Hastings Symmetric Update}
\usage{
update_metropolis(pname, loglik_func, logprior_func, parm_state, fixed,
  tune = 0.01, lower = NULL, upper = NULL)
}
\arguments{
\item{pname}{character string naming the parameter. The name must match something
in the parm_state list.}

\item{loglik_func, logprior_func}{log likelihood/prior function that takes the
parm_state and fixed lists as arguments.}

\item{parm_state}{list that names parameters and contains current values.}

\item{fixed}{list of named values that remain fixed in the model.}

\item{tune}{tuning parameter (standard deviation) in the normal proposal distribution.}

\item{lower, upper}{bounds for the parameter to restrict proposed values. Invalid
values are reflected back into the domain. If both bounds are provided, make
sure the tune is small so that reflections won't also be invalid.}
}
\description{
Function that performs a M-H update for a named parameter given the RAMSES 
structural elements and loglikelihood/logprior functions. Candidate values 
are drawn from a normal distribution, but it does allow for upper and lower bounds.
}
